fix some lint issues

+++ dataprovider/user.go
@@ -637,7 +637,7 @@ func (u *User) getNotificationFieldsAsSlice(action string) []string {
 	return []string{action, u.Username,
 		strconv.FormatInt(u.ID, 10),
 		strconv.FormatInt(int64(u.Status), 10),
-		strconv.FormatInt(int64(u.ExpirationDate), 10),
+		strconv.FormatInt(u.ExpirationDate, 10),
 		u.HomeDir,
 		strconv.FormatInt(int64(u.UID), 10),
 		strconv.FormatInt(int64(u.GID), 10),
